# bruno/scaffolder/examples/sample-swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Event API",
        "version": "1.0.0",
        "description": "Sample API for demonstrating Bruno scaffolder"
    },
    "servers": [
        {
            "url": "https://api-dev.contoso.com/api/v1",
            "description": "Development server"
        }
    ],
    "paths": {
        "/events/{eventId}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Event Detail",
                "operationId": "GetEventDetail",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique identifier of the event"
                    },
                    {
                        "name": "includeDetails",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "description": "Whether to include detailed information"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details retrieved successfully"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/events": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create Event",
                "operationId": "CreateEvent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully"
                    }
                }
            }
        },
        "/shareholders/lite": {
            "get": {
                "tags": [
                    "Shareholders"
                ],
                "summary": "Get Shareholders Lite",
                "operationId": "GetShareholdersLite",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Bearer token for authentication"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Page number for pagination"
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Number of items per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shareholders list retrieved successfully"
                    }
                }
            }
        },
        "/shareholders/{shareholderId}": {
            "put": {
                "tags": [
                    "Shareholders"
                ],
                "summary": "Update Shareholder",
                "operationId": "UpdateShareholder",
                "parameters": [
                    {
                        "name": "shareholderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique identifier of the shareholder"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Bearer token for authentication"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shareholder updated successfully"
                    },
                    "404": {
                        "description": "Shareholder not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Shareholders"
                ],
                "summary": "Patch Shareholder",
                "operationId": "UpdateShareholder",
                "parameters": [
                    {
                        "name": "shareholderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique identifier of the shareholder"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shareholder patched successfully"
                    }
                }
            }
        },
        "/shareholders/import": {
            "post": {
                "tags": [
                    "Shareholders"
                ],
                "summary": "Import Shareholders",
                "operationId": "ImportShareholders",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ShareholderImportItem"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shareholders imported successfully"
                    }
                }
            }
        },
        "/registrations": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create Registration",
                "operationId": "CreateRegistration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/RegistrationBase"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "promoCode": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration created successfully"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "Service is healthy"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Address": {
                "type": "object",
                "properties": {
                    "street": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "postCode": {
                        "type": "string"
                    }
                }
            },
            "CreateEventRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "isPublic": {
                        "type": "boolean"
                    },
                    "venue": {
                        "$ref": "#/components/schemas/Address"
                    }
                }
            },
            "ShareholderImportItem": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "shares": {
                        "type": "integer"
                    }
                }
            },
            "RegistrationBase": {
                "type": "object",
                "properties": {
                    "eventId": {
                        "type": "string"
                    },
                    "attendeeName": {
                        "type": "string"
                    },
                    "attendeeEmail": {
                        "type": "string"
                    }
                }
            }
        }
    }
}